backend/requirements.txt
concrete-ml>=1.9.0
numpy>=1.24.0
scikit-learn>=1.3.0

